Removed old AI and old files. Added a new generic random AI. game no longer deals with translation.

+++ framework/src/main/java/org/toop/framework/gameFramework/model/game/BoardProvider.java
@@ -1,5 +1,5 @@
 package org.toop.framework.gameFramework.model.game;
 
 public interface BoardProvider {
-    int[] getBoard();
+    long[] getBoard();
 }
