+++ framework/src/main/java/org/toop/framework/gameFramework/model/player/MoveProvider.java
@@ -3,5 +3,5 @@ package org.toop.framework.gameFramework.model.player;
 import org.toop.framework.gameFramework.model.game.TurnBasedGame;
 
 public interface MoveProvider<T extends TurnBasedGame<T>> {
-    int getMove(T game);
+    long getMove(T game);
 }
